requirements.txt
fastapi==0.111.0
gunicorn==20.1.0
requests==2.32.3
pydantic==2.7.4
python-decouple==3.8
SQLAlchemy==2.0.15
GeoAlchemy2==0.15.1
psycopg2-binary==2.9.6